+++ mini-kodein-android-compose/build.gradle
@@ -1,8 +1,6 @@
 /*
  * Copyright 2021 HyperDevs
  *
- * Copyright 2020 BQ
- *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -36,27 +34,22 @@ android {
     buildTypes {
         release {
             minifyEnabled false
-            proguardFiles getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro"
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
-
     compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_1_8
-        targetCompatibility = JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
-
     kotlinOptions {
-        jvmTarget = "1.8"
+        jvmTarget = '1.8'
     }
 }
 
 dependencies {
-    api project(':mini-kodein')
+    api project(':mini-kodein-android')
 
-    api "androidx.fragment:fragment-ktx:$fragment_version"
-    api "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_version"
-    api "org.kodein.di:kodein-di-framework-android-x:$kodein_version"
-    api "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version"
+    api "androidx.navigation:navigation-compose:2.4.0-alpha03"
 
     testImplementation "junit:junit:$junit_version"
     androidTestImplementation "androidx.test:runner:$test_runner_version"
