+++ mini-kodein-android-compose/mini-kodein-android.pro
@@ -1,0 +1,2 @@
+# Needed for injection
+-keep class * extends androidx.lifecycle.ViewModel { *; }
