+++ settings.gradle
@@ -15,5 +15,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-include ':app', ':mini-processor', ':mini-common', ':mini-android', ':mini-processor-test', 'mini-kodein', 'mini-kodein-android', 'mini-testing'
+include ':app', ':mini-processor', ':mini-common', ':mini-android', ':mini-processor-test', 'mini-kodein', 'mini-kodein-android', ':mini-kodein-android-compose', 'mini-testing'
